Refactor(loot): renaming - Names now reflect behavior (choosing items). No logic change.
