+++ src/main/java/org/example/service/loot/LootTable.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import org.example.entities.items.Item;
 
 /** Produces random loot items. Map/code using this doesn't know specific classes. */
-public interface LootFactory {
+public interface LootTable {
     /** One roll of the loot table. Empty = no drop. */
     Optional<Item> roll();
 }
